Add unambiguous window selection to tmux-launch
Previously, matching was performed on the window name only.  Add support
for an optional window ID prefix in the match string.  The string should
be of the form '[window_id:]window_name'.  If a window ID and a window
name is given, select the uniquely matching window, if it exists.  If
only a window name is given, select the first matching window.  If only
a window name is given but no matching window is found, the argument is
assumed to be an executable and invoked in a new window.

Examples:

tmux-launch @23:emacs   # select window via unique id:name
tmux-launch emacs -nw   # select first window matching 'emacs'
tmux-launch emacs -nw   # if no 'emacs' window exists, run 'emacs -nw'