Add Openbox root menu key binding to the desktop
Previously, the keyboard-mouse binding used to show the root menu only
worked in the "Frame" context.  The "Frame" context applies to all
application windows and their windows decorations, if any, but not the
desktop window.  Add this same binding to the "Desktop" context so that
it works everywhere.

This is further to commit e8b4028c5a.