Fix bindings on Bitstream Vera font substitutions
Alias elements have a binding of "weak" by default.  Consequently, if
the Bitstream Vera fonts are present, the preferred fonts will not be
substituted in their place.

Fontconfig author Keith Packard notes:

"The <prefer> binding is generally used when resolving the generic
family names (sans-serif, serif, monospace) and hence the names added
should be 'weak' so that they don't end up overriding the 'real' names
provided by applications."

In this case, the Bitstream Vera fonts are not simply generic names but
rather actual fonts being targeted for substitution.  Set the binding to
"same" such that the substitutions are actually performed.