Quote trailing backslash in .tmux.conf
According to the tmux(1) manpage in tmux version 2.2:

"Multiple commands may be specified together as part of a command
sequence.  Each command should be separated by spaces and a semicolon;
commands are executed sequentially from left to right and lines ending
with a backslash continue on to the next line, except when escaped by
another backslash.  A literal semicolon may be included by escaping it
with a backslash (for example, when specifying a command sequence to
bind-key)."

Lines beginning with '#' are treated as comments; however, the above
rules concerning trailing backslashes appear to apply to comment lines
as well.  A comment line ending in a backslash will continue onto the
following line which will also be treated as a comment.  Enclose the
trailing backslash in single quotes in order to avoid such behavior.