Fix 'cg' Bash alias to avoid cd to home directory
Fix the 'cg' Bash alias to avoid changing to the user's home directory
when called outside of a Git working tree.

Note that this case is not a no-op as the effective command of 'cd ./'
overwrites OLDPWD.  Also, the condition returns an exit status of 0
although a non-zero status would probably be more informative.