Make tmux-neww-here work in a remote tmux session
Previously, tmux-neww-here would fail if called outside of an X
environment.  One such scenario would be when connected to a remote tmux
session via SSH.  This was due to its dependcy upon tmux-launch which
requires an X display to be present.

Make tmux-neww-here open a new window in the current tmux session
instead of session 0 if invoked outside of an X display environment.