Rewrite 'cg' Bash alias as a Bash function
The 'cg' Bash alias overwrites OLDPWD even when called from the
top-level directory of a Git repository.  Reimplement 'cg' as a Bash
function such that calling it from the top-level directory of a Git
repository does not overwrite OLDPWD.  Additionally, make 'cg' return
the exit status of the underlying git rev-parse command when called
outside of a Git repository.