Fix GUI Emacs window width
Previously, if a window of 80 columns was desired and the fringes were
disabled, the width in the emacs.geometry X resource had be to set to
77.  This may have been due to a bug in Emacs.  At some point this bug
appears to have been fixed and emacs.geometry is now correctly
interpreted.  Fix the emacs.geometry X resource in order to restore the
GUI Emacs window to a width of 80 characters.