Emulate tmux v2.5 choose-window with choose-tree
As of tmux version 2.6, choose-window is aliased to 'choose-tree -w'.
The choose-tree command displays a list of windows in all sessions
whereas the choose-window command in tmux version 2.5 only listed
windows in the current session.

Replace the choose-window key bindings with a choose-tree command that
displays only the windows in the current session.  This change
approximates the behavior of choose-window in tmux version 2.5.  Unlike
choose-window, the choose-tree command still displays an additional line
item at the top of the window list which represents root of the tree and
the session as a whole.