Set tmux base-index option to 1
Set the tmux base-index option from 0 (the default) to 1 such that the
child window indices line up with the tree mode item indices as invoked
by the choose-tree command introduced in commit 6bebe20d24.

The choose-tree command displays an item at the top of the list which
represents the root of the tree and the session as a whole.  This
top-most item has an index of 0; thus, the indices of its children begin
at 1.  With the base-index option set to 1, the window indices will also
start at 1 instead of 0.

This is to make the tree mode less visually confusing and also make the
tree mode 0 to 9 key bindings more consistent with the default mode
0 to 9 window selection key bindings.