Fix urxvt console SVG icon
The edges of the rendered SVG icon are clipped in librsvg 2.44.x.  The
filter region was not computed correctly in librsvg 2.42.x.

By default, the <filter> element uses filterUnits="objectBoundingBox",
which constrains the filter region to the bounding box of the unstroked
object.

To produce the desired effect, add the attribute
filterUnits="userSpaceOnUse" to the <filter> element.  By default, this
causes the filter region to span the whole image, plus a little more.