Update alias to remote-viewer .bashrc
In virt-viewer version 3.0, the background color of the main window was
set to black.  However, more recent versions of virt-viewer refer to the
GTK theme when setting the background color of the main window.  Use the
GTK_THEME environment variable when invoking remote-viewer to explicitly
use the dark variant of the Adwaita GTK theme.