Fix detaching from a grouped tmux session
If the session is part of a session group, kill the session instead of
detaching it unless the session is the only remaining group member.

Previously, the 'detach' key binding would only check to see if the
session was part of a group but did not also check if it was the last
remaining session group member.  Detaching the last remaining session
group member would inadvertently kill the session.