Generalize the tmux detaching logic
If the session name matches the session group name, always detach it
even if there are other sessions in the group.

Previously, the detach key binding only guaranteed that the session
would be preserved if the session name was '0'.  The secondary console
session named 'A' was not considered.  If the 'A' session was killed
with other sessions attached to the 'A' session group, ambiguities would
arise when a new, ungrouped session also named 'A' was later created.