Unlink file only if it refers to the correct inode
Add a check to the 'dotfiles' script's 'unlink' subcommand to ensure the
file to be unlinked actually refers to the same inode as the target file.

Previously, if a file of the same name as the target existed in the
destination directory, it would be deleted regardless of whether or not
it actually pointed to the target.